Message class updates to more closely match iOS
