+++ code/messaging/src/test/java/com/adobe/marketing/mobile/MessageTests.java
@@ -20,6 +20,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import android.app.Activity;
+import android.content.Context;
+import android.webkit.ValueCallback;
+import android.webkit.WebSettings;
+import android.webkit.WebView;
 
 import com.adobe.marketing.mobile.services.ServiceProvider;
 import com.adobe.marketing.mobile.services.ui.AEPMessage;
@@ -31,10 +35,12 @@ import com.adobe.marketing.mobile.services.ui.MessageSettings;
 import com.adobe.marketing.mobile.services.ui.UIService;
 import com.adobe.marketing.mobile.services.ui.internal.MessagesMonitor;
 
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
+import org.mockito.ArgumentMatchers;
 import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -77,6 +83,10 @@ public class MessageTests {
     MessagingInternal mockMessagingInternal;
     @Mock
     ServiceProvider mockServiceProvider;
+    @Mock
+    WebView mockWebView;
+    @Mock
+    WebSettings mockWebSettings;
     @Captor
     ArgumentCaptor<MessagingEdgeEventType> messagingEdgeEventTypeArgumentCaptor;
     @Captor
@@ -86,7 +96,7 @@ public class MessageTests {
     private EventHub eventHub;
 
     @Before
-    public void setup() {
+    public void setup() throws Exception {
         eventHub = new EventHub("testEventHub", mockPlatformServices);
         mockCore.eventHub = eventHub;
 
@@ -100,7 +110,7 @@ public class MessageTests {
         }
     }
 
-    void setupMocks() {
+    void setupMocks() throws Exception {
         PowerMockito.mockStatic(MobileCore.class);
         PowerMockito.mockStatic(Event.class);
         PowerMockito.mockStatic(App.class);
@@ -114,6 +124,9 @@ public class MessageTests {
         Mockito.when(mockAEPMessage.getSettings()).thenReturn(mockAEPMessageSettings);
         Mockito.when(mockAEPMessageSettings.getParent()).thenReturn(mockMessage);
         Mockito.when(mockMessagingState.getExperienceEventDatasetId()).thenReturn("datasetId");
+        mockWebView = PowerMockito.mock(WebView.class);
+        PowerMockito.whenNew(WebView.class).withAnyArguments().thenReturn(mockWebView);
+        when(mockWebView.getSettings()).thenReturn(mockWebSettings);
     }
 
     Map<String, Object> setupXdmMap() {
@@ -244,9 +257,9 @@ public class MessageTests {
     }
 
     @Test
-    public void test_messageDismiss() {
+    public void test_messageDismiss_suppressAutoTrackFalse() {
         // test
-        message.dismiss();
+        message.dismiss(false);
 
         // verify aepMessage dismiss called
         verify(mockAEPMessage, times(1)).dismiss();
@@ -257,6 +270,18 @@ public class MessageTests {
         String interaction = interactionArgumentCaptor.getValue();
         assertEquals(eventType, MessagingEdgeEventType.IN_APP_DISMISS);
         assertEquals(null, interaction);
+    }
+
+    @Test
+    public void test_messageDismiss_suppressAutoTrackTrue() {
+        // test
+        message.dismiss(true);
+
+        // verify aepMessage dismiss called
+        verify(mockAEPMessage, times(1)).dismiss();
+
+        // verify no dismissed tracking event
+        verify(mockMessagingInternal, times(0)).handleInAppTrackingInfo(messagingEdgeEventTypeArgumentCaptor.capture(), interactionArgumentCaptor.capture(), any(Message.class));
     }
 
     @Test
@@ -292,8 +317,8 @@ public class MessageTests {
         // test
         message.overrideUrlLoad(mockAEPMessage, "adbinapp://dismiss?interaction=deeplinkclicked&link=https://adobe.com");
 
-        // expect 2 events: deeplink click tracking + dismissed tracking
-        verify(mockMessagingInternal, times(2)).handleInAppTrackingInfo(messagingEdgeEventTypeArgumentCaptor.capture(), interactionArgumentCaptor.capture(), any(Message.class));
+        // expect 1 event: deeplink click tracking
+        verify(mockMessagingInternal, times(1)).handleInAppTrackingInfo(messagingEdgeEventTypeArgumentCaptor.capture(), interactionArgumentCaptor.capture(), any(Message.class));
         List<MessagingEdgeEventType> capturedEvents = messagingEdgeEventTypeArgumentCaptor.getAllValues();
         List<String> capturedInteractions = interactionArgumentCaptor.getAllValues();
         // verify interact tracking event
@@ -301,12 +326,6 @@ public class MessageTests {
         String interaction = capturedInteractions.get(0);
         assertEquals(MessagingEdgeEventType.IN_APP_INTERACT, displayTrackingEvent);
         assertEquals("deeplinkclicked", interaction);
-
-        // verify custom delegate tracking event
-        MessagingEdgeEventType dismissTrackingEvent = capturedEvents.get(1);
-        interaction = capturedInteractions.get(1);
-        assertEquals(MessagingEdgeEventType.IN_APP_DISMISS, dismissTrackingEvent);
-        assertEquals(null, interaction);
 
         // verify showUrl called
         verify(mockUIService, times(1)).showUrl("https://adobe.com");
@@ -350,6 +369,61 @@ public class MessageTests {
         // verify no tracking event
         verify(mockMessagingInternal, times(0)).handleInAppTrackingInfo(messagingEdgeEventTypeArgumentCaptor.capture(), interactionArgumentCaptor.capture(), any(Message.class));
 
+    }
+
+    @Test
+    public void test_loadJavascript_WithScriptHandlersSet() {
+        // setup
+        AdobeCallback<String> callback = new AdobeCallback<String>() {
+            @Override
+            public void call(String s) {
+                Assert.assertEquals("hello world", s);
+            }
+        };
+        // set js webview
+        message.view = mockWebView;
+        // set scriptHandlers map
+        Map<String, WebViewJavascriptInterface> scriptHandlers = new HashMap<>();
+        scriptHandlers.put("test", new WebViewJavascriptInterface(callback));
+        Whitebox.setInternalState(message, "scriptHandlers", scriptHandlers);
+
+        // test
+        message.evaluateJavascript("(function test(hello world) { return(arg); })()");
+        // verify evaluate javascript called
+        verify(mockWebView, times(1)).evaluateJavascript(ArgumentMatchers.contains("(function test(hello world) { return(arg); })()"), any(ValueCallback.class));
+    }
+
+    @Test
+    public void test_loadJavascript_WithNoScriptHandlersSet() {
+        // set js webview
+        message.view = mockWebView;
+        // test
+        message.evaluateJavascript("(function test(hello world) { return(arg); })()");
+
+        // verify evaluate javascript called
+        verify(mockWebView, times(0)).evaluateJavascript(ArgumentMatchers.contains("(function test(hello world) { return(arg); })()"), any(ValueCallback.class));
+    }
+
+    @Test
+    public void test_loadJavascriptWhenJavascriptIsNull() {
+        // setup
+        AdobeCallback<String> callback = new AdobeCallback<String>() {
+            @Override
+            public void call(String s) {
+                Assert.assertEquals("hello world", s);
+            }
+        };
+        // set js webview
+        message.view = mockWebView;
+        // set scriptHandlers map
+        Map<String, WebViewJavascriptInterface> scriptHandlers = new HashMap<>();
+        scriptHandlers.put("test", new WebViewJavascriptInterface(callback));
+        Whitebox.setInternalState(message, "scriptHandlers", scriptHandlers);
+        // test
+        message.evaluateJavascript(null);
+
+        // verify evaluate javascript not called
+        verify(mockWebView, times(0)).evaluateJavascript(anyString(), any(ValueCallback.class));
     }
 
     class CustomMessageDelegate extends MessageDelegate {
